requirements (5).txt
transformers==4.40.1
torch
gradio==4.14.0
pdfplumber==0.10.3
python-docx==1.1.0
